+++ NEWS.md
@@ -1,4 +1,4 @@
-## version 0.1.0.99
+## version 0.1.1
 
 ### Features
 
@@ -6,6 +6,11 @@
     * with / without dependencies (pkgs in `Suggests` field)
 * add refresh button to reload the app
 * add local and remote versions from DESCRIPTION file and last commit date (#2)
+* add popup on mouse over events:
+    + column 1: package name (repo), popup: "user/repo"
+    + column 2: local version (DESCRIPTION), popup: "ref@/sha1"
+    + column 3: remote version (DESCRIPTION), popup: "ref@/sha1"
+    
 
 ### Bug fix